+++ features/notifications/index.ts
@@ -2,3 +2,4 @@ export type { NotificationItem } from './notification.types';
 export { useNotifications } from './hooks/use-notifications.query';
 export { useNotificationActions } from './hooks/use-notification-actions';
 export { formatRelativeTime } from './lib/format-relative-time';
+export { NotificationBell } from './components/notification-bell';
