--- shared/components/notification/index.ts
@@ -1,4 +1,0 @@
-export { BellFilled } from './bell-filled';
-export { NotificationBell } from './notification-bell';
-export { NotificationDropdownContent } from './notification-dropdown-content';
-export { NotificationItem } from './notification-item';
